test(buzz-sdk): split broker/tests.rs into six concern modules
`crates/buzz-sdk/src/broker/tests.rs` is 2,638 lines against the repo's
1,000-line ceiling, and the next broker test addition would push it further.
Split along the `// ── Section ──` headers the file already carried, so the cut
lines are the ones the author already drew.

| module | lines | holds |
|---|---|---|
| `tests/mod.rs` | 184 | shared fixtures, helpers, `mod` declarations |
| `tests/schema.rs` | 689 | action coverage, envelope round-trip and rejection |
| `tests/wire_schema.rs` | 361 | exact-key-set tables, the no-secret invariant |
| `tests/validation.rs` | 714 | argument bounds, read provenance, result/status tables |
| `tests/identities.rs` | 423 | one spelling per identity; the result type's absent reader |
| `tests/client.rs` | 312 | identical-bytes retry, client trait, test doubles |

Six modules, not the four #6746 sketched and not the five of the previous
revision. Two sections each force their own module for the same reason: the
wire-schema section would leave `schema.rs` at 1,045 lines, and the
identity-canonicalization section added by #6742's review revision would leave
`validation.rs` at roughly 1,130. Either would be a split that still violates
the cap it exists to satisfy. Splitting them out keeps every file under the
ceiling without moving a single test across a concern boundary.

Verified as a pure move rather than asserted. Concatenating the six modules in
the original section order and stripping only the added scaffolding — module
docs, five `mod` declarations, per-file `use super::*` — reproduces the original
file exactly, with a single deliberate exception: `member_paths` becomes
`pub(super)` because it is now shared across two modules. That one-token
visibility change is the whole diff of the reconstruction:

    -fn member_paths(value: &serde_json::Value, prefix: &str, out: &mut Vec<String>) {
    +pub(super) fn member_paths(value: &serde_json::Value, prefix: &str, out: &mut Vec<String>) {

And the test set, not merely its size: the same 37 broker tests exist before and
after, compared with module paths normalized away, so a rename plus a deletion
could not hide behind a matching count.

Validation: `cargo test -p buzz-sdk --all-targets` 299 passed / 0 failed,
identical to the baseline at 7c853e0fb; package clippy `-D warnings` exit 0 with
no dead-code warning from the moved helper; `cargo fmt --all --check` clean;
`cargo doc` warning-free; every file under 1,000 lines.

Closes #6746. Stacked on #6742. Contract context: #6467.
